+++ src/test/java/br/com/os/dao/UsuarioDAOTest.java
@@ -173,7 +173,7 @@ public class UsuarioDAOTest {
 	@Ignore
 	public void deletar() {
 	//TESTADO - OK	
-		Integer cod = 2;
+		Long cod = 2L;
 		UsuarioDAO usuarioDAO = new UsuarioDAO();
 		Usuario udeletado = new Usuario();
 		udeletado.setNome("rer");
